+++ micro_manager/adaptivity/adaptivity.py
@@ -5,13 +5,20 @@ from math import exp
 from typing import Callable
 import importlib
 from micro_manager.tools.logging_wrapper import Logger
+from micro_manager.config import Config
+from micro_manager.micro_simulation import MicroSimulationClass
 
 import numpy as np
 
 
 class AdaptivityCalculator:
     def __init__(
-        self, configurator, nsims, micro_problem_cls, base_logger, rank
+        self,
+        configurator: Config,
+        nsims: int,
+        micro_problem_cls: MicroSimulationClass,
+        base_logger: Logger,
+        rank: int,
     ) -> None:
         """
         Class constructor.